This is the requirements file.
You want to have a requirements file where all external 3rd party apps are listed and pinned to their versions. That way, anyone willing to use your project can do it by using exactly the same libraries and versions you used. In time, the number of libraries and/or their versions may change. That's why this file mus be updated in a regular basis. https://www.easydevmixin.com/posts/2015/06/creating-a-requirements-file/