Remove unused code.
I know, I know... What if I want to reuse it later? Yeah, yeah... OK, truth is when you comment or stop using code, the chances to use it again are les than 1%, so why keep it? That said, if you ever want to reuse it, that's why we're using git and keep a tracking history, right?