Big guns out. Restructuring code.
You should aim to make your code expansible. The refactoring in this commit is all about it. I've taken out the code that processes the model and put it into another module.
What this means, is that now you have the ability to create new models and use them, no need to delete the old one from coronavirus10.py. Now you can have as many as you want.
These are the important things you should know:
  - there is a new package called . It's a package because it has a  in it.
  - in the package there is a module called . In there you define a class (follow the example). The important thing here is that you give the class the config you want to use and the data. See how it is done in the code. Another important thing is that all model classes should implement a  method. If you do that, it will be super easy to run any model you want. In Python, this is called a protocol https://www.python.org/dev/peps/pep-0544/\#terminology.