Main point of entrance.
Now the script is an app. It has a main point of entrance where you pass the data and can give a config. To execute it, the command is:
python coronavirus19.py <path to data file> [--config <path to config file>]
Where:
  - <path to data file> is a mandatory argument. It is the path to your data.
  - --config <path to config file> is optional. If given, the config file is a JSON file that defines all configuration options. If not given, it will use the default configuration.

You can also run:

python coronavirus19.py --help

To see the help menu.

Enjoy